fix float32 to float32

+++ predictor_life_simple/dataloader.py
@@ -20,7 +20,7 @@ class LifeGameDataset(Dataset):
         file_path = self.file_list[file_idx]
         data = np.load(file_path)  # Shape: [T, N, N]
         t = np.random.randint(0, data.shape[0] - 2)
-        x, x_1, x_2 = map(lambda x: torch.tensor(x, dtype=torch.float64), [data[t], data[t + 1], data[t + 2]])
+        x, x_1, x_2 = map(lambda x: torch.tensor(x, dtype=torch.float32), [data[t], data[t + 1], data[t + 2]])
         x = (torch.stack([x, x_1], dim=0) / 255)
         y = (torch.stack([x_1, x_2], dim=0) / 255)
         return x, y
